+++ lib/src/data_transfer_objects/has_balance/has_balance_result.g.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: package_api_docs, non_constant_identifier_names
+
+part of 'has_balance_result.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$HasBalanceResultImpl _$$HasBalanceResultImplFromJson(
+        Map<String, dynamic> json) =>
+    _$HasBalanceResultImpl(
+      exist: json['exist'] as bool,
+    );
+
+Map<String, dynamic> _$$HasBalanceResultImplToJson(
+        _$HasBalanceResultImpl instance) =>
+    <String, dynamic>{
+      'exist': instance.exist,
+    };